Adding Git Ignore, Fixed xcode warnings
Adding git ignore file to remove tracking of files that shouldn't be tracked in Github.
Some of these files include cocoapods and other xcode files that constantly change in development, and makes pull requests less cluttered.
Xcode was complaining of a missing file, added back to the project
Xcode had a warning that was resolved when removing .plist from copy bundle resources