Camera Refactoring
Noticed that the methods used to take a photo were deprecated.
It's always best to update those methods because in the next release, they may not be usable anymore.
Removed a bit of extra work that was being done during every image capture.
Cleaned up code a bit and added guards to avoid possible app crashing due to force unwrapping.